Update: 認証リクエスト処理をAppコンポーネントに移動
