+++ frontend/src/pages/App.jsx
@@ -1,11 +1,29 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import Top from "./Top";
 import Workout from "./Workout";
-import { useState } from "react";
+import { useCallback, useEffect, useState } from "react";
 import AuthContext from "../Contexts/AuthContext";
+import { getUser } from "../utils/auth";
+import { isEmptyObj } from "../utils/objectControl";
 
 function App() {
   const [authInfo, setAuthInfo] = useState({})
+
+  // ユーザーの認証情報を取得
+  const firstSetUserInfo = useCallback(
+    async () => {
+      const res = await getUser();
+      setAuthInfo(res.data)
+    }, [setAuthInfo]
+  )
+
+  useEffect(() => {
+    console.log(authInfo);
+    
+    if(isEmptyObj(authInfo)) {
+      firstSetUserInfo();
+    }
+  }, [authInfo, firstSetUserInfo])
 
   return (
     <>
