+++ frontend/src/pages/Top.jsx
@@ -3,59 +3,39 @@
 // - 各運動と必要な運動時間をRails apiから取得して表示
 // - 各運動データのボタンから、運動管理画面に遷移(state: {体重, 摂取カロリー, 運動データ})
 
-import { useCallback, useContext, useEffect, useState } from 'react'
+import { useEffect, useState } from 'react'
 import '../builds/build.css'
 import axiosCunstom from '../utils/axiosCustoms';
 import WorkOutCard from '../components/top/WorkOutCard';
 import CalorieForm from '../components/top/CalorieForm';
 import { useLocation } from 'react-router-dom';
 import SideMenu from '../components/general/SideMenu';
-import { getUser } from '../utils/auth';
-import { isEmptyObj } from '../utils/objectControl';
-import AuthContext from '../Contexts/AuthContext';
 
 function Top() {
   // ページ遷移時の処理
   const location = useLocation();
-  const {authInfo, setAuthInfo} = useContext(AuthContext);
 
   const [weight        , setWeight        ] = useState(location.state?.weight || 50);
   const [intakedCalorie, setIntakedCalorie] = useState(location.state?.intakedCalorie || 0);
   const [workoutsObj   , setWorkoutsObj   ] = useState([]);
 
   //クリック時に、apiから運動データを取得する処理
-  const fetchWorkoutsData = useCallback(
-    async () => {
-      const res = await axiosCunstom.get("/work_outs", {
-        params: {
-          weight:      weight,
-          kcal_intake: intakedCalorie,
-        }
-      });
-      setWorkoutsObj(res.data);
-    }, [intakedCalorie, weight]
-  )
-
-  const firstAuthInfo = useCallback(
-    async () => {
-      const res = await getUser();
-      setAuthInfo(res.data)
-    }, [setAuthInfo]
-  );
-
-  // ユーザーの認証情報を取得
-  useEffect(() => {
-    if(isEmptyObj(authInfo)) {
-      firstAuthInfo();
-    }
-  }, [authInfo, firstAuthInfo])
+  const fetchWorkoutsData = async () => {
+    const res = await axiosCunstom.get("/work_outs", {
+      params: {
+        weight:      weight,
+        kcal_intake: intakedCalorie,
+      }
+    });
+    setWorkoutsObj(res.data);
+  }
 
   // ページ遷移した時、遷移元から体重・カロリーを受け取っているならfetchWorkoutsData
   useEffect(() => {
     if (location.state?.weight && location.state?.intakedCalorie) {
       fetchWorkoutsData();
     }
-  }, [fetchWorkoutsData, location])
+  }, [location]) //fetchWorkoutsData()の更新に反応すると、フォーム入力値が変更される(setWeight)たびにGETリクエストが発行される
 
   return (
     <>
